backend/pyproject.toml
[project]
name = "zhomind-backend"
version = "0.1.0"
description = "ZhoMind backend foundation"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.116.0",
  "uvicorn[standard]>=0.35.0",
  "pydantic-settings>=2.10.0",
  "python-multipart>=0.0.20",
  "sqlalchemy>=2.0.0",
  "alembic>=1.13.0",
  "aiosqlite>=0.20.0",
]

[dependency-groups]
dev = [
  "pytest>=8.4.0",
  "httpx>=0.28.0",
  "ruff>=0.12.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.ruff]
line-length = 150
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B"]
ignore = ["B008"]

[tool.ruff.format]
quote-style = "double"

[tool.uv]
package = false

[project.scripts]
dev = "uvicorn src.main:app --reload"
test = "pytest"
lint = "ruff check src tests"
format = "ruff format src tests"
